pyproject.toml
[tool.poetry]
name = "coin_monitor"
version = "0.1.0"
description = "Access to CoinGecko API endpoints"
authors = ["Fergus Strangways-Dixon <fergusdixon101@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "3.1"
djangorestframework = "^3.11.1"
asgiref = "^3.2.10"
python-dotenv = "^0.14.0"
pydantic = "^1.6.1"
markdown = "^3.2.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"
flake8 = "^3.8.3"
isort = "^5.4.2"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
